refactor(utils): move CreateMTDFile and DownloadFromMTDFile to their own respective files
