+++ src/Utils.js
@@ -235,83 +235,3 @@ export const RequestWithMeta = R.uncurryN(2, (HTTP) => R.compose(
   Rx.flatMap(RequestThread(HTTP)),
   FlattenMeta$
 ))
-export const DownloadFromMTDFile = ({FILE, HTTP, mtdPath}) => {
-  /**
-   * Open file to read+append
-   */
-  const fd$ = FILE.open(O.just([mtdPath, 'r+']))
-
-  /**
-   * Retrieve File size on disk
-   */
-  const size$ = LocalFileSize$({FILE, fd$})
-
-  /**
-   * Retrieve Meta info
-   */
-  const metaPosition$ = MetaPosition$({size$})
-  const meta$ = ReadJSON$({FILE, fd$, position$: metaPosition$})
-
-  /**
-   * Make a HTTP request for each thread
-   */
-  const {response$, buffer$} = demuxFPH(
-    ['buffer$', 'response$'], RequestWithMeta(HTTP, meta$)
-  )
-
-  /**
-   * Create write params and save buffer+offset to disk
-   */
-  const bufferWritten$ = WriteBuffer({FILE, fd$, buffer$})
-
-  /**
-   * Update META info
-   */
-  const nMeta$ = SetMetaOffsets({meta$, bufferWritten$})
-
-  /**
-   * Persist META to disk
-   */
-  const metaWritten$ = FILE.write(CreateWriteBufferAtParams({
-    fd$,
-    buffer$: JSToBuffer$(RxThrottleComplete(meta$.pluck('metaWrite'), nMeta$)),
-    position$: size$
-  }))
-
-  /**
-   * Create sink$
-   */
-  return mux({
-    metaWritten$, response$,
-    localFileSize$: size$,
-    fdR$: fd$, metaPosition$,
-    meta$: O.merge(nMeta$, meta$)
-  })
-}
-export const CreateMTDFile = ({FILE, HTTP, options}) => {
-  /**
-   * Create a new file
-   */
-  const fd$ = FILE.open(O.just([options.mtdPath, 'w']))
-
-  /**
-   * Retreive file size on remote server
-   */
-  const size$ = RemoteFileSize$({HTTP, options})
-
-  /**
-   * Create initial meta data
-   */
-  const meta$ = CreateMeta$({options, size$})
-
-  /**
-   * Create a new file with meta info appended at the end
-   */
-  const written$ = FILE.write(CreateWriteBufferAtParams({
-    FILE,
-    fd$: fd$,
-    buffer$: JSToBuffer$(meta$),
-    position$: size$
-  }))
-  return mux({written$, meta$, remoteFileSize$: size$, fdW$: fd$})
-}
